Don't make header media fixed in bg anymore.
Right now I can't get this to scroll nicely on iOS. Main culprit
is the use of fixed positioning on the header media. The more
important thing is stutter free scrolling on mobile, so just
make image static in the flow for now.

Also made header and media same size. No need for them to differ
in size now.