Add drop shadow to logo, and remove header media gradient.
The box shadow gives the logo some depth and helps it stand out
from the header media.

I just personally don't like a gradient on the header media, so
I removed it. Furthermore, it made it very hard to predict the
appearance of an image, since the gradient darkens the image.
I prefer a WYSIWYG experience, so what I see in Photoshop is what
I get on the final site.